@@ -2,7 +2,7 @@
 
 open System
 open System.Data
-open System.Data.SqlClient
+open Microsoft.Data.SqlClient
 open System.Collections.Generic
 open FSharp.Data.SqlClient.Internals
 
