@@ -22,7 +22,7 @@ open System
 open System.Text
 open System.Data
 open System.Collections.Generic
-open System.Data.SqlClient
+open Microsoft.Data.SqlClient
 open System.Runtime.Serialization
 open System.Runtime.Serialization.Json
 open System.IO
